+++ cnccode/functions/Common.inc.php
@@ -2,8 +2,14 @@
 /**
  * determine whether location is an internal stock location
  */
-define('COMMON_MYSQL_DATETIME', 'Y-m-d H:i:s');
-define('COMMON_MYSQL_DATE', 'Y-m-d');
+define(
+    'COMMON_MYSQL_DATETIME',
+    'Y-m-d H:i:s'
+);
+define(
+    'COMMON_MYSQL_DATE',
+    'Y-m-d'
+);
 
 function common_isAnInternalStockLocation($customerID)
 {
@@ -34,25 +40,49 @@ function common_isAnInternalStockSupplier($supplierID)
 function common_convertDateDMYToYMD($dateDMY)
 {
     if ($dateDMY != '') {
-        $dateArray = explode('/', $dateDMY);
-        return ($dateArray[2] . '-' . str_pad($dateArray[1], 2, '0', STR_PAD_LEFT) . '-' . str_pad($dateArray[0], 2, '0', STR_PAD_LEFT));
+        $dateArray = explode(
+            '/',
+            $dateDMY
+        );
+        return ($dateArray[2] . '-' . str_pad(
+                $dateArray[1],
+                2,
+                '0',
+                STR_PAD_LEFT
+            ) . '-' . str_pad(
+                $dateArray[0],
+                2,
+                '0',
+                STR_PAD_LEFT
+            ));
     } else {
         return '';
     }
 }
 
-function common_dateFormat($timestamp, $format = 1)
+function common_dateFormat($timestamp,
+                           $format = 1
+)
 {
     $timestamp = trim($timestamp);
     // only do something if a date is passed in
     if ($timestamp > 0):
         // check if date contains '/' char i.e. from user entered data dd/mm/yyyy
-        if (strstr($timestamp, "/")):
-            $parts = explode("/", $timestamp);
+        if (strstr(
+            $timestamp,
+            "/"
+        )):
+            $parts = explode(
+                "/",
+                $timestamp
+            );
             // $parts[0] = day
             // $parts[1] = month
             // $parts[2] = year
-            $timestamp = date("Y-m-d H:i:s", strtotime("$parts[2]-$parts[1]-$parts[0]"));
+            $timestamp = date(
+                "Y-m-d H:i:s",
+                strtotime("$parts[2]-$parts[1]-$parts[0]")
+            );
         endif;
         // convert to unix time stamp for easier formatting
         $timestamp = strtotime($timestamp);
@@ -60,11 +90,17 @@ function common_dateFormat($timestamp, $format = 1)
         switch ($format):
             case 1:
                 // 25/12/2003
-                $rDate = date("d/m/Y", $timestamp);
+                $rDate = date(
+                    "d/m/Y",
+                    $timestamp
+                );
                 break;
             case 2:
                 // 25/12/2003 12:00:00
-                $rDate = date("d/m/Y H:i:s", $timestamp);
+                $rDate = date(
+                    "d/m/Y H:i:s",
+                    $timestamp
+                );
                 break;
         endswitch;
         return $rDate;
@@ -77,8 +113,16 @@ function common_dateFormat($timestamp, $format = 1)
  */
 function common_convertHHMMToDecimal($hhMM)
 {
-    $hours = substr($hhMM, 0, 2);
-    $minutes = substr($hhMM, 3, 2);
+    $hours = substr(
+        $hhMM,
+        0,
+        2
+    );
+    $minutes = substr(
+        $hhMM,
+        3,
+        2
+    );
 
     $minutesAsFraction = $minutes / 60;
 
@@ -87,17 +131,26 @@ function common_convertHHMMToDecimal($hhMM)
 
 function convertHHMMToMinutes($hhMM)
 {
-    $time = explode(':', $hhMM);
+    $time = explode(
+        ':',
+        $hhMM
+    );
     return ($time[0] * 60.0 + $time[1] * 1.0);
 }
 
 if (!function_exists('str_split')) {
-    function str_split($string, $split_length = 1)
+    function str_split($string,
+                       $split_length = 1
+    )
     {
         $strlen = strlen($string);
 
         for ($i = 0; $i < $strlen; $i += $split_length) {
-            $array[] = substr($string, $i, $split_length);
+            $array[] = substr(
+                $string,
+                $i,
+                $split_length
+            );
         }
 
         return $array;
@@ -108,13 +161,26 @@ function common_convertDecimalToHHMM($decimalTime)
 {
 
     $hoursInteger = floor($decimalTime);
-    $hoursString = str_pad($hoursInteger, 2, '0', STR_PAD_LEFT);
-    $minutesString = str_pad(floor(60 * ($decimalTime - $hoursInteger)), 2, '0', STR_PAD_LEFT);
+    $hoursString = str_pad(
+        $hoursInteger,
+        2,
+        '0',
+        STR_PAD_LEFT
+    );
+    $minutesString = str_pad(
+        floor(60 * ($decimalTime - $hoursInteger)),
+        2,
+        '0',
+        STR_PAD_LEFT
+    );
     return $hoursString . ':' . $minutesString;
 
 }
 
-function common_inRange($value, $startRange, $endRange)
+function common_inRange($value,
+                        $startRange,
+                        $endRange
+)
 {
     if (($value > $endRange) OR ($value < $startRange)) {
         return FALSE;
@@ -123,15 +189,22 @@ function common_inRange($value, $startRange, $endRange)
     }
 }
 
-function common_numberFormat($number, $fuzz = 0.00000000001)
+function common_numberFormat($number,
+                             $fuzz = 0.00000000001
+)
 {
-    return sprintf("%.2f", (($number >= 0) ? ($number + $fuzz) : ($number - $fuzz)));
+    return sprintf(
+        "%.2f",
+        (($number >= 0) ? ($number + $fuzz) : ($number - $fuzz))
+    );
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // convert a string to hex of ascii code - blocks email spiders
 //////////////////////////////////////////////////////////////////////////////////////////
-function strtohex($string, $prefix = "%")
+function strtohex($string,
+                  $prefix = "%"
+)
 {
     for ($i = 0; $i < strlen($string); $i++) {
         $return .= $prefix . dechex(ord($string[$i]));
@@ -139,7 +212,9 @@ function strtohex($string, $prefix = "%")
     return $return;
 }
 
-function common_getHTMLEmailFooter($senderName, $senderEmail)
+function common_getHTMLEmailFooter($senderName,
+                                   $senderEmail
+)
 {
 
     // do not apply a footer - CNC mail server does this now
@@ -168,11 +243,31 @@ function common_roundUpToQuarter($value)
     return $value;
 }
 
-function common_dateDiffMins($startDate, $startTime)
+function common_dateDiffMins($startDate,
+                             $startTime
+)
 {
-    $startDate = explode('-', $startDate);
-    $startTime = explode(':', $startTime);
-    $mins = abs(bcdiv(time() - mktime($startTime[0], $startTime[1], 0, $startDate[1], $startDate[2], $startDate[0]), 60));
+    $startDate = explode(
+        '-',
+        $startDate
+    );
+    $startTime = explode(
+        ':',
+        $startTime
+    );
+    $mins = abs(
+        bcdiv(
+            time() - mktime(
+                $startTime[0],
+                $startTime[1],
+                0,
+                $startDate[1],
+                $startDate[2],
+                $startDate[0]
+            ),
+            60
+        )
+    );
     return $mins;
 }
 
@@ -193,13 +288,19 @@ function common_getNextWorkingDay()
 {
     $nextdays = array(strtotime('+1 day'), strtotime('+2 days'), strtotime('+3 days'));
     for ($i = 0; $i < count($nextdays); $i++) {
-        $daynum = (int)date('w', $nextdays[$i]);
+        $daynum = (int)date(
+            'w',
+            $nextdays[$i]
+        );
         if (($daynum > 0) && ($daynum < 6)) {
             $nextDate = $nextdays[$i];
             break;
         }
     }
-    return date('d/m/Y', $nextDate);
+    return date(
+        'd/m/Y',
+        $nextDate
+    );
 }
 
 /**
@@ -210,42 +311,160 @@ function common_getNextWorkingDay()
  */
 function common_stripEverything($description)
 {
-    $description = str_replace("\r\n", '', trim($description));
-    $description = str_replace("\r", '', trim($description));
-    $description = str_replace("\n", '', $description);
-    $description = str_replace("\t", '', $description);
-    $description = str_replace('<br />', "", $description);
-    $description = str_replace('<br/>', "", $description);
-    $description = str_replace('<BR/>', "", $description);
-    $description = str_replace('<BR>', "", $description);
-    $description = str_replace('<p>', "", $description);
-    $description = str_replace('</p>', "", $description);
-    $description = str_replace('<P>', "", $description);
-    $description = str_replace('</P>', "", $description);
-    $description = str_replace('&nbsp;', " ", $description);
-    $description = str_replace('&quot;', "'", $description);
+    $description = str_replace(
+        "\r\n",
+        '',
+        trim($description)
+    );
+    $description = str_replace(
+        "\r",
+        '',
+        trim($description)
+    );
+    $description = str_replace(
+        "\n",
+        '',
+        $description
+    );
+    $description = str_replace(
+        "\t",
+        '',
+        $description
+    );
+    $description = str_replace(
+        '<br />',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '<br/>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '<BR/>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '<BR>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '<p>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '</p>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '<P>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '</P>',
+        "",
+        $description
+    );
+    $description = str_replace(
+        '&nbsp;',
+        " ",
+        $description
+    );
+    $description = str_replace(
+        '&quot;',
+        "'",
+        $description
+    );
     $description = strip_tags($description);
     $description = trim($description);
     return $description;
 
 }
 
+function getExpiryDate(DateTime $installDate,
+                       DateTime $today = null
+)
+{
+
+    if ($today == null) {
+        $today = new DateTime();
+    }
+
+    //get next expiry date
+
+    $expiryDay = (int)$installDate->format('d');
+    $expiryMonth = (int)$installDate->format('m');
+
+    // we need to check
+    $expiryYear = (int)$today->format('Y');
+
+    if ($expiryMonth < (int)$today->format('m') ||
+        $expiryMonth == (int)$today->format('m') &&
+        $expiryDay < (int)$today->format('d')) {
+        $expiryYear += 1;
+    }
+
+    $nextExpiryDate = DateTime::createFromFormat(
+        'Y-m-d',
+        "$expiryYear-$expiryMonth-$expiryDay"
+    );
+
+    $difference = (int)$nextExpiryDate->diff($today)->format('%m');
+
+    $expiryDate = clone $nextExpiryDate;
+    if ($difference < 3) {
+        $expiryDate->add(new DateInterval("P1Y"));
+    }
+
+    return $expiryDate;
+}
+
 function common_getUKBankHolidays($year)
 {
     $utFirstJan = strtotime('1st jan ' . $year);
 
-    $firstJanDay = date('N', $utFirstJan);
+    $firstJanDay = date(
+        'N',
+        $utFirstJan
+    );
 
     if ($firstJanDay > 5) { // sat or sun
-        $holidays[] = date('Y-m-d', strtotime('first monday of january ' . $year));
+        $holidays[] = date(
+            'Y-m-d',
+            strtotime('first monday of january ' . $year)
+        );
     } else {
-        $holidays[] = date('Y-m-d', $utFirstJan);
+        $holidays[] = date(
+            'Y-m-d',
+            $utFirstJan
+        );
     }
 
     $utEasterSunday = easter_date($year);
-    $holidays[] = date('Y-m-d', strtotime('last friday', $utEasterSunday));
-    $holidays[] = date('Y-m-d', strtotime('next monday', $utEasterSunday));
-    $holidays[] = date('Y-m-d', strtotime('first monday of may ' . $year));
+    $holidays[] = date(
+        'Y-m-d',
+        strtotime(
+            'last friday',
+            $utEasterSunday
+        )
+    );
+    $holidays[] = date(
+        'Y-m-d',
+        strtotime(
+            'next monday',
+            $utEasterSunday
+        )
+    );
+    $holidays[] = date(
+        'Y-m-d',
+        strtotime('first monday of may ' . $year)
+    );
 
     /*
     Example of adhock holiday adding
@@ -256,22 +475,67 @@ function common_getUKBankHolidays($year)
       $holidays[] = '2012-06-05';
     }
     */
-    $holidays[] = date('Y-m-d', strtotime('last monday of may ' . $year));
+    $holidays[] = date(
+        'Y-m-d',
+        strtotime('last monday of may ' . $year)
+    );
 
-    $holidays[] = date('Y-m-d', strtotime('last monday of august ' . $year));  // August bh
+    $holidays[] = date(
+        'Y-m-d',
+        strtotime('last monday of august ' . $year)
+    );  // August bh
 
     $xmasDay = date(strtotime('25th december ' . $year));
-    if (date('N', $xmasDay) == 5) { // falls on friday
-        $holidays[] = date('Y-m-d', $xmasDay);
-        $holidays[] = date('Y-m-d', strtotime('next monday', $xmasDay));
+    if (date(
+            'N',
+            $xmasDay
+        ) == 5) { // falls on friday
+        $holidays[] = date(
+            'Y-m-d',
+            $xmasDay
+        );
+        $holidays[] = date(
+            'Y-m-d',
+            strtotime(
+                'next monday',
+                $xmasDay
+            )
+        );
     }
-    if (date('N', $xmasDay) > 5) { // falls on sat or sun
-        $holidays[] = date('Y-m-d', strtotime('next monday', $xmasDay));
-        $holidays[] = date('Y-m-d', strtotime('next tuesday', $xmasDay));
+    if (date(
+            'N',
+            $xmasDay
+        ) > 5) { // falls on sat or sun
+        $holidays[] = date(
+            'Y-m-d',
+            strtotime(
+                'next monday',
+                $xmasDay
+            )
+        );
+        $holidays[] = date(
+            'Y-m-d',
+            strtotime(
+                'next tuesday',
+                $xmasDay
+            )
+        );
     }
-    if (date('N', $xmasDay) < 5) { // falls on mon to thurs
-        $holidays[] = date('Y-m-d', $xmasDay);
-        $holidays[] = date('Y-m-d', strtotime('next day', $xmasDay));
+    if (date(
+            'N',
+            $xmasDay
+        ) < 5) { // falls on mon to thurs
+        $holidays[] = date(
+            'Y-m-d',
+            $xmasDay
+        );
+        $holidays[] = date(
+            'Y-m-d',
+            strtotime(
+                'next day',
+                $xmasDay
+            )
+        );
     }
 
     return $holidays;